refactor: split up subtree.rs, renamed to element
