+++ grovedb/src/element/constructor.rs
@@ -1,0 +1,111 @@
+use crate::{Element, ElementFlags};
+use crate::element::{MaxReferenceHop, SumValue};
+use crate::reference_path::ReferencePathType;
+
+impl Element {
+    #[cfg(feature = "full")]
+    // TODO: improve API to avoid creation of Tree elements with uncertain state
+    pub fn empty_tree() -> Self {
+        Element::new_tree(Default::default())
+    }
+
+    #[cfg(feature = "full")]
+    pub fn empty_tree_with_flags(flags: Option<ElementFlags>) -> Self {
+        Element::new_tree_with_flags(Default::default(), flags)
+    }
+
+    #[cfg(feature = "full")]
+    pub fn empty_sum_tree() -> Self {
+        Element::new_sum_tree(Default::default())
+    }
+
+    #[cfg(feature = "full")]
+    pub fn empty_sum_tree_with_flags(flags: Option<ElementFlags>) -> Self {
+        Element::new_sum_tree_with_flags(Default::default(), flags)
+    }
+
+    #[cfg(feature = "full")]
+    pub fn new_item(item_value: Vec<u8>) -> Self {
+        Element::Item(item_value, None)
+    }
+
+    #[cfg(feature = "full")]
+    pub fn new_item_with_flags(item_value: Vec<u8>, flags: Option<ElementFlags>) -> Self {
+        Element::Item(item_value, flags)
+    }
+
+    #[cfg(feature = "full")]
+    pub fn new_sum_item(value: i64) -> Self {
+        Element::SumItem(value, None)
+    }
+
+    #[cfg(feature = "full")]
+    pub fn new_sum_item_with_flags(value: i64, flags: Option<ElementFlags>) -> Self {
+        Element::SumItem(value, flags)
+    }
+
+    #[cfg(feature = "full")]
+    pub fn new_reference(reference_path: ReferencePathType) -> Self {
+        Element::Reference(reference_path, None, None)
+    }
+
+    #[cfg(feature = "full")]
+    pub fn new_reference_with_flags(
+        reference_path: ReferencePathType,
+        flags: Option<ElementFlags>,
+    ) -> Self {
+        Element::Reference(reference_path, None, flags)
+    }
+
+    #[cfg(feature = "full")]
+    pub fn new_reference_with_hops(
+        reference_path: ReferencePathType,
+        max_reference_hop: MaxReferenceHop,
+    ) -> Self {
+        Element::Reference(reference_path, max_reference_hop, None)
+    }
+
+    #[cfg(feature = "full")]
+    pub fn new_reference_with_max_hops_and_flags(
+        reference_path: ReferencePathType,
+        max_reference_hop: MaxReferenceHop,
+        flags: Option<ElementFlags>,
+    ) -> Self {
+        Element::Reference(reference_path, max_reference_hop, flags)
+    }
+
+    #[cfg(feature = "full")]
+    pub fn new_tree(maybe_root_key: Option<Vec<u8>>) -> Self {
+        Element::Tree(maybe_root_key, None)
+    }
+
+    #[cfg(feature = "full")]
+    pub fn new_tree_with_flags(
+        maybe_root_key: Option<Vec<u8>>,
+        flags: Option<ElementFlags>,
+    ) -> Self {
+        Element::Tree(maybe_root_key, flags)
+    }
+
+    #[cfg(feature = "full")]
+    pub fn new_sum_tree(maybe_root_key: Option<Vec<u8>>) -> Self {
+        Element::SumTree(maybe_root_key, 0, None)
+    }
+
+    #[cfg(feature = "full")]
+    pub fn new_sum_tree_with_flags(
+        maybe_root_key: Option<Vec<u8>>,
+        flags: Option<ElementFlags>,
+    ) -> Self {
+        Element::SumTree(maybe_root_key, 0, flags)
+    }
+
+    #[cfg(feature = "full")]
+    pub fn new_sum_tree_with_flags_and_sum_value(
+        maybe_root_key: Option<Vec<u8>>,
+        sum_value: SumValue,
+        flags: Option<ElementFlags>,
+    ) -> Self {
+        Element::SumTree(maybe_root_key, sum_value, flags)
+    }
+}
