+++ grovedb/src/estimated_costs/worst_case_costs.rs
@@ -23,7 +23,7 @@ use storage::{worst_case_costs::WorstKeyLength, Storage};
 #[cfg(feature = "full")]
 use crate::{
     batch::{key_info::KeyInfo, KeyInfoPath},
-    subtree::{SUM_ITEM_COST_SIZE, SUM_TREE_COST_SIZE, TREE_COST_SIZE},
+    element::{SUM_ITEM_COST_SIZE, SUM_TREE_COST_SIZE, TREE_COST_SIZE},
     Element, ElementFlags, Error, GroveDb,
 };
