+++ grovedb/src/operations/proof/generate.rs
@@ -4,20 +4,21 @@ use costs::{
 };
 #[cfg(feature = "full")]
 use merk::{
-    proofs::{encode_into, Node, Op},
-    tree::value_hash,
-    KVIterator, Merk, ProofWithoutEncodingResult,
+    KVIterator,
+    Merk,
+    proofs::{encode_into, Node, Op}, ProofWithoutEncodingResult, tree::value_hash,
 };
 #[cfg(feature = "full")]
 use storage::{rocksdb_storage::PrefixedRocksDbStorageContext, StorageContext};
 
 #[cfg(feature = "full")]
 use crate::{
-    operations::proof::util::{write_to_vec, ProofType, EMPTY_TREE_HASH},
-    reference_path::path_from_reference_path_type,
-    subtree::raw_decode,
-    Element, Error, GroveDb, PathQuery, Query,
+    Element,
+    Error,
+    GroveDb, operations::proof::util::{EMPTY_TREE_HASH, ProofType, write_to_vec}, PathQuery, Query, reference_path::path_from_reference_path_type,
 };
+#[cfg(feature = "full")]
+use crate::element::helpers::raw_decode;
 
 #[cfg(feature = "full")]
 type LimitOffset = (Option<u16>, Option<u16>);
